requirements.txt
pandas
numpy
matplotlib
seaborn
streamlit
boto3
opencv-python
ultralytics
flask
flask-cors
easyocr
pyYAML
python-box
joblib
dvc
scikit-learn
mlflow
-e .